#234 Moved text style model from core to api so that is can be used from custom components.
