+++ pramen/api/src/main/scala/za/co/absa/pramen/api/notification/TableHeader.scala
@@ -14,20 +14,6 @@
  * limitations under the License.
  */
 
-package za.co.absa.pramen.core.notify.message
+package za.co.absa.pramen.api.notification
 
-sealed trait Style
-
-object Style {
-  case object Normal extends Style
-
-  case object Bold extends Style
-
-  case object Success extends Style
-
-  case object Warning extends Style
-
-  case object Exception extends Style
-
-  case object Error extends Style
-}
+case class TableHeader(textElement: TextElement, align: Align = Align.Left)
