+++ pramen/core/src/main/scala/za/co/absa/pramen/core/notify/message/TableBuilderHtml.scala
@@ -16,6 +16,8 @@
 
 package za.co.absa.pramen.core.notify.message
 
+import za.co.absa.pramen.api.notification.{Align, Style, TableHeader, TextElement}
+
 import scala.collection.mutable.ListBuffer
 
 class TableBuilderHtml extends TableBuilder {
